+++ src/UplinkrServiceProvider.php
@@ -21,6 +21,7 @@ use Uplinkr\Console\Commands\Prune\PruneStorageCommand;
 use Uplinkr\Console\Commands\UplinkrConfigCommand;
 use Uplinkr\Console\Commands\UplinkrIamAliveCommand;
 use Uplinkr\Console\Commands\UplinkrInstallCommand;
+use Uplinkr\Console\Commands\UplinkrSettingsCommand;
 use Uplinkr\Handler\Probe\ResultHandler;
 use Uplinkr\Handler\Probe\UrlHandler;
 use Uplinkr\Handler\Project\Probes\ProbeAllProjectsHandler;
@@ -31,6 +32,8 @@ use Uplinkr\Notifications\Channels\UplinkrWebhookChannel;
 use Uplinkr\Objects\Config\UplinkrConfig;
 use Uplinkr\Storage\FileProbeResultsStorage;
 use Uplinkr\Storage\FileProjectStorage;
+use Uplinkr\Storage\FileSettingsStorage;
+use Uplinkr\Support\Logger;
 use Uplinkr\Support\Sanitizer;
 
 /**
@@ -76,6 +79,7 @@ class UplinkrServiceProvider extends ServiceProvider
                 UplinkrInstallCommand::class,
                 UplinkrConfigCommand::class,
                 UplinkrIamAliveCommand::class,
+                UplinkrSettingsCommand::class,
             ]);
 
             $this->app->booted(function () {
@@ -87,7 +91,6 @@ class UplinkrServiceProvider extends ServiceProvider
                 $schedule = app(Schedule::class);
                 $probeCron = (string)config('uplinkr.scheduler.cron');
                 $alertCron = config('uplinkr.scheduler.alert_cron');
-                $statusInterval = (int)config('uplinkr.scheduler.status_interval', 0);
                 $alertCron = is_string($alertCron) && $alertCron !== '' ? $alertCron : $probeCron;
 
                 $schedule->command('uplinkr:project:run-probes --force')
@@ -99,10 +102,25 @@ class UplinkrServiceProvider extends ServiceProvider
                     ->cron($alertCron)
                     ->withoutOverlapping();
 
-                if ($statusInterval >= 1 && $statusInterval <= 59) {
-                    $schedule->command('uplinkr:iam-alive')
-                        ->cron(sprintf('*/%d * * * *', $statusInterval))
-                        ->withoutOverlapping();
+                try {
+                    $settingsStorage = app(FileSettingsStorage::class);
+                    $iamAlive = $settingsStorage->getIamAliveSettings();
+                    $enabled = (bool)($iamAlive['enabled'] ?? false);
+                    $intervalHours = (int)($iamAlive['interval_hours'] ?? 24);
+
+                    if ($enabled && $intervalHours >= 1 && $intervalHours <= 24) {
+                        $cron = $intervalHours === 24
+                            ? '0 0 * * *'
+                            : sprintf('0 */%d * * *', $intervalHours);
+
+                        $schedule->command('uplinkr:iam-alive')
+                            ->cron($cron)
+                            ->withoutOverlapping();
+                    }
+                } catch (\JsonException $exception) {
+                    Logger::log()->warning('Unable to load uplinkr settings.json for scheduler.', [
+                        'reason' => $exception->getMessage(),
+                    ]);
                 }
             });
         }
@@ -166,6 +184,12 @@ class UplinkrServiceProvider extends ServiceProvider
             return new ProbeAllProjectsHandler(
                 $app->make(ProjectStorageInterface::class),
                 $app->make(UrlHandler::class)
+            );
+        });
+
+        $this->app->singleton(FileSettingsStorage::class, function ($app) {
+            return new FileSettingsStorage(
+                $app->make(UplinkrConfig::class)
             );
         });